chore(db): Use flask-sqlalchemy to decrease code boilerplate
