+++ src/annotation/.env.example
@@ -1,4 +1,4 @@
-FLASK_APP=app.py
+# FLASK_APP=app.py
 # Use it for login; issue your own google credentials
 GOOGLE_CLIENT_ID=1-2-3
 GOOGLE_CLIENT_SECRET=4-5-6
