+++ src/annotation/annotations_app/config.py
@@ -13,6 +13,8 @@ class Config:
     # comma-separated list of emails; these emails will be considered superusers on each login
     AUTH_WHITELISTED_EMAILS = (os.environ.get("AUTH_WHITELISTED_EMAILS") or "").split(",")
 
+    SQLALCHEMY_DATABASE_URI = os.environ["SQLALCHEMY_URL"]
+
 
 LOGGING_CONFIG = {
     'version': 1,
