+++ src/annotation/annotations_app/views/users.py
@@ -1,7 +1,8 @@
 from flask import abort, jsonify, request
 from marshmallow import ValidationError
 
-from annotations_app import app, schemas
+from annotations_app.flask_app import app
+from annotations_app import schemas
 from annotations_app.models.base import User
 from annotations_app.utils import db_session, superuser_required
